email-automation/mcp-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
redis==5.0.1
openai==1.3.0
python-multipart==0.0.6